# openapi.yaml
openapi: 3.0.3
info:
  title: Clean Data Hub Enterprise Ingestion API
  version: 1.1.0
  description: |
    API specifications for the Clean Data Hub platform. Includes robust document ingestion, 
    refinement pipelines, structured metadata generation, custom PII redaction gates, 
    SAML SSO flows, billing checks, and tenant operations.
servers:
  - url: http://localhost:3000
    description: Development Local server
paths:
...
  /api/documents:
    get:
      summary: Get all documents
      tags:
        - Documents
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Scoped documents list
    post:
      summary: Ingest raw text document into pipeline
      tags:
        - Documents
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - rawContent
              properties:
                name:
                  type: string
                rawContent:
                  type: string
                type:
                  type: string
                  enum: [PDF, DOCX, XLSX, TXT]
                  default: TXT
                size:
                  type: string
                connector:
                  type: string
      responses:
        '201':
          description: Created
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT